add Mybatis Generator code

+++ lineage-server/lineage-server-start/src/main/resources/sql-script/mysql/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE `bas_task`
     `modify_user_id` bigint(20)   NOT NULL DEFAULT '0',
     `ctime`          bigint(20)   NOT NULL,
     `mtime`          bigint(20)   NOT NULL,
-    `invalid`        tinyint(3)   NOT NULL DEFAULT '0',
+    `invalid`        int(3)       NOT NULL DEFAULT '0',
     PRIMARY KEY (`task_id`),
     KEY `task_name_idx` (`task_name`)
 ) ENGINE = InnoDB
@@ -48,7 +48,7 @@ CREATE TABLE `rel_task_lineage`
     `target_table`    varchar(128) NOT NULL,
     `target_column`   varchar(128) NOT NULL,
     `transform`       varchar(128) NOT NULL,
-    `invalid`         tinyint(3)   NOT NULL DEFAULT '0',
+    `invalid`         int(3)       NOT NULL DEFAULT '0',
     PRIMARY KEY (`rid`),
     KEY `task_id_idx` (`task_id`)
 ) ENGINE = InnoDB
@@ -65,7 +65,7 @@ CREATE TABLE `bas_plugin`
     `modify_user_id` bigint(20)   NOT NULL DEFAULT '0',
     `ctime`          bigint(20)   NOT NULL,
     `mtime`          bigint(20)   NOT NULL,
-    `invalid`        tinyint(3)   NOT NULL DEFAULT '0',
+    `invalid`        int(3)       NOT NULL DEFAULT '0',
     PRIMARY KEY (`plugin_id`),
     KEY `plugin_name_idx` (`plugin_name`)
 ) ENGINE = InnoDB
@@ -84,7 +84,7 @@ CREATE TABLE `bas_catalog`
     `modify_user_id`   bigint(20)   NOT NULL DEFAULT '0',
     `ctime`            bigint(20)   NOT NULL,
     `mtime`            bigint(20)   NOT NULL,
-    `invalid`          tinyint(3)   NOT NULL DEFAULT '0',
+    `invalid`          int(3)       NOT NULL DEFAULT '0',
     PRIMARY KEY (`catalog_id`),
     KEY `catalog_name_idx` (`catalog_name`)
 ) ENGINE = InnoDB
@@ -95,16 +95,16 @@ DROP TABLE IF EXISTS `bas_function`;
 CREATE TABLE `bas_function`
 (
     `function_id`     bigint(20) AUTO_INCREMENT,
-    `function_name`   varchar(128)        NOT NULL,
-    `function_format` varchar(128)        NOT NULL,
-    `function_path`   varchar(1024)       NOT NULL,
-    `function_class`  varchar(128)        NOT NULL,
+    `function_name`   varchar(128)  NOT NULL,
+    `function_format` varchar(128)  NOT NULL,
+    `function_path`   varchar(1024) NOT NULL,
+    `function_class`  varchar(128)  NOT NULL,
     `descr`           varchar(255),
-    `create_user_id`  bigint(20)          NOT NULL DEFAULT '0',
-    `modify_user_id`  bigint(20)          NOT NULL DEFAULT '0',
-    `ctime`           bigint(20)          NOT NULL,
-    `mtime`           bigint(20)          NOT NULL,
-    `invalid`         tinyint(3) unsigned NOT NULL DEFAULT '0',
+    `create_user_id`  bigint(20)    NOT NULL DEFAULT '0',
+    `modify_user_id`  bigint(20)    NOT NULL DEFAULT '0',
+    `ctime`           bigint(20)    NOT NULL,
+    `mtime`           bigint(20)    NOT NULL,
+    `invalid`         int(3)        NOT NULL DEFAULT '0',
     PRIMARY KEY (`function_id`)
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8;
